style: Dto 컨벤션 적용
