+++ src/main/java/underdogs/devbie/notice/dto/NoticeDescriptionResponse.java
@@ -4,23 +4,28 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import underdogs.devbie.notice.domain.Language;
 import underdogs.devbie.notice.domain.NoticeDescription;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Builder
 @Getter
 public class NoticeDescriptionResponse {
 
     private Set<String> languages;
     private String content;
 
-    public NoticeDescriptionResponse(NoticeDescription noticeDescription) {
-        this.languages = noticeDescription.getLanguages()
+    public static NoticeDescriptionResponse from(NoticeDescription noticeDescription) {
+        Set<String> languages = noticeDescription.getLanguages()
             .stream()
             .map(Language::getName)
             .collect(Collectors.toSet());
-        this.content = noticeDescription.getContent();
+
+        return new NoticeDescriptionResponse(languages, noticeDescription.getContent());
     }
 }
