+++ src/main/java/underdogs/devbie/notice/dto/NoticeResponses.java
@@ -1,21 +1,44 @@
 package underdogs.devbie.notice.dto;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import underdogs.devbie.notice.domain.Language;
+import underdogs.devbie.notice.domain.Notice;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Builder
 @Getter
 public class NoticeResponses {
 
     private List<NoticeResponse> noticeResponses;
 
-    public static NoticeResponses from(List<NoticeResponse> noticeResponses) {
-        return new NoticeResponses(new ArrayList<>(noticeResponses));
+    public static NoticeResponses from(List<Notice> notices) {
+        List<NoticeResponse> response = notices.stream()
+            .map(notice -> NoticeResponse.builder()
+                .id(notice.getId())
+                .name(notice.getCompany().getName())
+                .image(notice.getImage())
+                .languages(collectLanguageName(notice))
+                .jobPosition(notice.getJobPosition())
+                .build())
+            .collect(Collectors.toList());
+
+        return new NoticeResponses(response);
+    }
+
+    private static Set<String> collectLanguageName(Notice notice) {
+        return notice.getNoticeDescription()
+            .getLanguages()
+            .stream()
+            .map(Language::getName)
+            .collect(Collectors.toSet());
     }
 }
